common accuracy/threshhold for detector and recogniser

+++ utilities/facemanagement/facepipeline_p.h
@@ -241,7 +241,7 @@ protected:
 
     ThumbnailImageCatcher*          catcher;
     KFaceIface::RecognitionDatabase database;
-    double                          recognitionThreshold;
+    float                          recognitionThreshold;
     FacePipeline::Private* const    d;
 };
